requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
dnspython==2.8.0
et_xmlfile==2.0.0
fastapi==0.121.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
motor==3.7.1
openpyxl==3.1.5
passlib==1.7.4
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.5
python-dotenv==1.2.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
